addressed comments by chkoar
