+++ sklearn/feature_selection/__init__.py
@@ -19,6 +19,7 @@ from .variance_threshold import VarianceThreshold
 
 from .rfe import RFE
 from .rfe import RFECV
+
 from .sequential_feature_selector import SequentialFeatureSelector
 
 from .from_model import SelectFromModel
